requirements.txt
yfinance
pandas
numpy
matplotlib
scikit-learn
tensorflow
keras
ta
backtrader
gym
stable-baselines3
torch
alpaca-trade-api
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
schedule
